requirements.txt
# Core MCP dependencies
fastmcp==2.6.1
mcp==1.9.2

# Jira integration
jira==3.8.0

# Configuration and utilities
python-dotenv==1.1.0
PyYAML==6.0.2

# HTTP requests (used by jira library)
requests==2.32.3
urllib3==2.4.0

# Async HTTP client for Slack API
httpx>=0.28.1

# Type hints and validation
typing_extensions==4.14.0

# Performance monitoring (optional)
# rich==14.1.0  # For better console output
